% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{bir_file}
\alias{bir_file}
\title{Get path to bir_to example}
\usage{
bir_file(path = NULL)
}
\arguments{
\item{path}{Name of file. If \code{NULL}, the example files will be listed.}
}
\description{
bir_to comes bundled with some example files in its \code{inst/extdata}
directory. This function make them easy to access.
}
\examples{
bir_file()
bir_file("example.xlsx")
}
